Move states mixin to separate directory.
